+++ app/src/main/kotlin/com/platform/openemoji/screens/EmojiDetailScreen.kt
@@ -1,6 +1,5 @@
 package com.platform.openemoji.screens
 
-import android.app.Activity
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
@@ -11,8 +10,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
@@ -22,10 +19,9 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.core.text.HtmlCompat
 import androidx.navigation.NavController
-import com.google.android.gms.ads.interstitial.InterstitialAd
 import com.platform.openemoji.ads.AdSettings
 import com.platform.openemoji.ads.InlineAd
-import com.platform.openemoji.ads.InterstitialAds
+import com.platform.openemoji.ads.InterstitialAd
 import com.platform.openemoji.emoji.Emoji
 import com.platform.openemoji.emoji.IconCopy
 import com.platform.openemoji.favorites.FavoritesViewModel
@@ -42,12 +38,9 @@ fun EmojiDetailScreen(
 
     // Start loading an interstitial fullscreen ad. Only if this ad is loaded
     // by the time the user presses the return arrow, will the ad be shown.
-    val interstitialAd = remember { mutableStateOf<InterstitialAd?>(null) }
     if (AdSettings.get().displayInterstitialAdFromEmojiDetailScreen) {
         LaunchedEffect(LocalLifecycleOwner.current) {
-            InterstitialAds.load(context) {
-                interstitialAd.value = it
-            }
+            InterstitialAd.load(context)
         }
     }
     Column(
@@ -63,7 +56,7 @@ fun EmojiDetailScreen(
             navController = navController,
             modifier = Modifier.testTag("emojiDetailBackButton"),
         ) {
-            interstitialAd.value?.show(context as Activity)
+            InterstitialAd.show(context)
         }
 
         InlineAd()
